requirements.txt
Flask==3.0.2
requests==2.31.0
pycryptodome==3.20.0
SQLAlchemy==2.0.30
redis==4.2.0
PyMySQL==1.1.1